tools/requirements.txt
# Deps for tools/ scripts only — not required to run the test suite.
openpyxl>=3.1
pyyaml>=6.0
